fix(email): pre-PR review fixes for mail-proxy integration
- Fix send_mail batch action to use getMailProxy(raise_if_missing=False)
  preventing crash when proxy service exists but is not fully configured
- Extract shared proxy dashboard code into MailProxyDashboard component
  eliminating ~200 lines of duplication between mp_controls and sending_dashboard
- Add missing trailing newline to mp_endpoint.py
